+++ CHANGELOG.md
@@ -12,6 +12,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - remove `MARK` from `Interface` trait
 - use function instead constant for type names
 - rename `GraphqlType` to `TypeName`
+- rename `InputType` to `InputTypeName`
+- rename `OutputType` to `OutputTypeName`
 
 ## [0.5.4] - 2023-01-30